requirements.txt
argon2==0.1.10
asgiref==3.2.10
attrs==20.3.0
awsebcli==3.19.2
bcrypt==3.1.7
blessed==1.17.11
botocore==1.19.23
cached-property==1.5.2
cement==2.8.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
cycler==0.10.0
Django==3.0.6
django-crispy-forms==1.9.2
django-environ==0.4.5
djangorestframework==3.11.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.8
img2pdf==0.3.3
importlib-metadata==2.0.0
jdcal==1.4
jmespath==0.10.0
joblib==0.14.0
jsonschema==3.2.0
kiwisolver==1.1.0
lxml==4.2.5
matplotlib==3.1.1
MouseInfo==0.0.4
paramiko==2.7.2
pathspec==0.5.9
pdf2image==1.6.0
pdfminer.six==20181108
pikepdf==1.6.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.19
pycryptodome==3.8.2
PyGetWindow==0.0.7
Pygments==2.6.1
PyMsgBox==1.0.7
PyNaCl==1.4.0
pyobjc==5.2
pyobjc-core==5.2

pyparsing==2.4.2
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.22
python-dateutil==2.8.0
python-docx==0.8.10
PyTweening==1.0.3
pytz==2019.2
PyYAML==5.3.1
reportlab==3.5.23
requests==2.22.0
semantic-version==2.5.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.2
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
tqdm==4.31.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
